Default Values

LightOSM.set_defaults — Function

LightOSM.set_defaults(kwargs...)

Sets default values that LightOSM uses when generating the graph. All arguments are optional.

Keyword Arguments

maxspeeds::AbstractDict{String,<:Real}: If no maxspeed way tag is available, these values are used instead based on the value of the highway way tag. If no highway way tag is available, the value for "other" is used. Unit is km/h. Default value: Dict( "motorway" => 100, "trunk" => 100, "primary" => 100, "secondary" => 100, "tertiary" => 50, "unclassified" => 50, "residential" => 50, "other" => 50 )

lanes::AbstractDict{String,<:Integer}: If no lanes way tag is available, these values are used instead based on the value of the highway way tag. If no highway way tag is available, the value for "other" is used. Default value: Dict( "motorway" => 3, "trunk" => 3, "primary" => 2, "secondary" => 2, "tertiary" => 1, "unclassified" => 1, "residential" => 1, "other" => 1 )

lane_efficiency::AbstractDict{<:Integer,<:Real}: Gives the lane efficiency based on number of lanes. 1.0 is used for any number of lanes not specified here. Default value: LANE_EFFICIENCY = Dict( 1 => 0.7, 2 => 0.8, 3 => 0.9, 4 => 1.0 )

building_height_per_level::Integer: If the height building tag is not available, it is calculated by multiplying this value by the number of levels from the building:levels tag. Unit is metres. Default value: 4

max_building_levels::Integer: If the building:levels tag is not available, a number is randomly chosen between 1 and this value. Default value: 3