GeoLocation Methods

LightOSM.distance — Function

distance(A::Union{Vector{GeoLocation}, GeoLocation, Vector{<:Node}, Node, Vector{<:AbstractFloat}},
         B::Union{Vector{GeoLocation}, GeoLocation, Vector{<:Node}, Node, Vector{<:AbstractFloat}},
         type::Symbol=:haversine
         )::Union{Vector{AbstractFloat}, AbstractFloat}

Calculates the distance (km) between two points or two vectors of points.

Arguments

A::Union{Vector{GeoLocation}, GeoLocation, Vector{<:Node}, Node, Vector{<:AbstractFloat}}: Vector of origin points.

B::Union{Vector{GeoLocation}, GeoLocation, Vector{<:Node}, Node, Vector{<:AbstractFloat}}: Vector of destination points.

method::Symbol=:haversine: Either :haversine or :euclidean.

Return

Union{Vector{AbstractFloat}, AbstractFloat}: Distance between origin and destination points in km.

LightOSM.heading — Function

heading(A::Vector{GeoLocation}, B::Vector{GeoLocation}, return_units::Symbol=:degrees)::Vector{AbstractFloat}

Calculates headings / bearings between a vector of origin GeoLocations A and vector of destination GeoLocations B. Depending on the return_units chosen, the return angle is in range of [-π, π] if :radians or [-180, 180] if :degrees. Additionally, adjusts destination longitude in case the straight line path between A and B crosses the International Date Line.

Arguments

A::Vector{GeoLocation}: Vector of origin GeoLocations.

B::Vector{GeoLocation}: Vector of destination GeoLocations.

return_units::Symbol=:degrees: Either :radians or :degrees.

Return

Vector{AbstractFloat}: Vector of headings / bearings in range of [-π, π] if :radians or [-180, 180] if :degrees.

LightOSM.calculate_location — Function

calculate_location(lats::Vector{<:Number},
                   lons::Vector{<:Number},
                   headings::Vector{<:Number},
                   distances::Vector{<:Number}
                   )::Vector{NamedTuple}

Calculates next location given origin Geolocations (lats, lons), headings (degrees) and distances (km).

Arguments

origins::Vector{GeoLocation}: Vector of origin GeoLocations (lats, lons).

headings::Vector{<:Number}: Vector of headings to next location (degrees).

distances::Vector{<:Number}: Vector of distances to next location (km).

Return

Vector{GeoLocation}: Vector of GeoLocations (lats, lons) of next locations.

LightOSM.to_matrix — Function

to_matrix(locations::Vector{GeoLocation})::Matrix{AbstractFloat}

Converts a vector of GeoLocations to a matrix where the columns are in order of: latitude, longitude, altitude.

LightOSM.to_cartesian — Function

to_cartesian(location::Vector{GeoLocation})::Matrix{<:AbstractFloat}

Converts a vector of GeoLocations to (x, y, z) cartesian coordinates (based on radius of the Earth).

LightOSM.bounding_box_from_point — Function

bounding_box_from_point(point::GeoLocation, radius::Number)::NamedTuple

Calculates the coordinates of the bounding box given a centroid point and radius (km).

Arguments

point::GeoLocation: Centroid of the bounding box as an GeoLocation.

radius::Number: Radius in km of the bounding box (to each corner).

Return

NamedTuple: Named tuple with attributes minlat, minlon, maxlat, right_lon.