Shortest Path

LightOSM.shortest_path — Function

shortest_path(g::OSMGraph,
              origin::U,
              destination::U;
              algorithm::Symbol=:dijkstra,
              save_dijkstra_state::Bool=false,
              heuristic::Union{Function,Nothing}=nothing
              )::Vector{U} where {U <: Integer}

Calculates the shortest path between two OpenStreetMap node ids.

Arguments

g::OSMGraph: Graph container.

origin::U: Origin OpenStreetMap node id.

destination::U: Destination OpenStreetMap node id.

algorithm::Symbol=:dijkstra: Shortest path algorithm, either. :dijkstra or :astar.

save_dijkstra_state::Bool=false: Option to cache dijkstra parent states from a single source.

heuristic::Union{Function,Nothing}=nothing: Option to use custom astar heuristic, default haversine distance will be used if left blank.

Return

Vector{U}: Array of OpenStreetMap node ids making up the shortest path.

Calculates the shortest path between two OpenStreetMap node objects.

LightOSM.weights_from_path — Function

weights_from_path(g::OSMGraph{U,T,W}, path::Vector{T})::Vector{W} where {U <: Integer,T <: Integer,W <: Real}

Extracts edge weights from a path using g.weights.

Arguments

g::OSMGraph: Graph container.

path::Vector{T}: Array of OpenStreetMap node ids.

Return

Vector{W}: Array of edge weights, distances are in km, time is in hours.

LightOSM.astar — Function

astar(g::AbstractGraph{U},
      src::W;
      goal::Union{W,Nothing}=nothing,
      distmx::AbstractMatrix{T}=weights(g),
      restrictions::Union{AbstractDict{V,Vector{MutableLinkedList{V}}},Nothing}=nothing,
      heuristic::Function=h(u, v) =  0.0,
      )::Vector{U} where {T <: Real,U <: Integer,V <: Integer,W <: Integer}

A* shortest path algorithm taken and adapted from LightGraphs.jl. This version improves runtime speed, memory usage, has a more flexible heruistic function, and accounts for OpenStreetMap turn restrictions.

Arguments

g::AbstractGraph{U}: LightGraphs abstract graph object.

src::W: Source vertex.

goal::Union{W,Nothing}=nothing: Optional target vertex as a break condition.

distmx::AbstractMatrix{T}=weights(g): Optional weight matrix.

restrictions::Union{AbstractDict{V,Vector{MutableLinkedList{V}}},Nothing}=nothing: Optional

dictionary of restrictions, keyed by vertex, each restriction is a linked list of vertices, any path containing the entire list will have an infinite cost assigned to it.

heuristic::Function=h(u, v) = 0.0: Heuristic cost function, takes a source and target vertex, default is 0.

Return

Vector{U}: Array parent veritces from which the shortest path can be extracted.

LightOSM.dijkstra — Function

Dijkstra shortest path algorithm, same as A* but without a heuristic cost function.