Nearest Node

LightOSM.nearest_node — Function

nearest_node(g::OSMGraph,
             points::Vector{GeoLocation},
             n_neighbours::Integer=1
             )::Tuple{Vector{Vector{Integer}}, Vector{Vector{AbstractFloat}}}

Finds nearest nodes from a vector of GeoLocations using a NearestNeighbors.jl KDTree.

Arguments

g::OSMGraph: Graph container.

points::Vector{GeoLocation}: Vector of query points.

n_neighbours::Integer: Number of neighbours to query for each point.

Return

Tuple of neighbours and straight line euclidean distances from each point ([[neighbours]...], [[dists]...]).

Finds nearest nodes from a vector of node ids, the query node id is excluded from neighbours.