Shortest Path

LightOSM.shortest_path — Function

shortest_path(g::OSMGraph,
              origin::U,
              destination::U;
              weights=g.weights,
              restrictions=g.indexed_restrictions,
              algorithm::Symbol=:dijkstra,
              save_dijkstra_state::Bool=false,
              heuristic::Union{Function,Nothing}=nothing
              )::Vector{U} where {U <: Integer}

Calculates the shortest path between two OpenStreetMap node ids.

Arguments

g::OSMGraph: Graph container.

origin: Origin OpenStreetMap node or node id.

destination: Destination OpenStreetMap node or node id.

weights: Matrix of node to node edge weights, defaults to g.weights. If a custom weights matrix is being used with algorithm set to :astar, make sure that a correct heuristic is being used.

algorithm::Symbol=:dijkstra: Shortest path algorithm, either. :dijkstra or :astar.

save_dijkstra_state::Bool=false: Option to cache dijkstra parent states from a single source.

restrictions, defaults to g.indexed_restrictions.

heuristic::Function=default_heuristic(g): Option to use custom astar heuristic. Defaults to a function h(u, v) -> haversine(u, v), i.e. returns the haversine distances between u is the current node and v is the neighbouring node. Haversine distance will be used if g.weight_type is :distance, otherwise an estimated time will be used by dividing the harversine distance by the speed at u (specified in the maxspeed tag).

cost_adjustment::Function=default_cost_adjustment(g): Turn restictions dictionary. Pass in nothing to not use

Return

Vector{Integer}: Array of OpenStreetMap node ids making up the shortest path.

LightOSM.weights_from_path — Function

weights_from_path(g::OSMGraph{U,T,W}, path::Vector{T}; weights=g.weights)::Vector{W} where {U <: Integer,T <: Integer,W <: Real}

Extracts edge weights from a path using the weight matrix stored in g.weights unless a different matrix is passed to the weights kwarg.

Arguments

g::OSMGraph: Graph container.

path::Vector{T}: Array of OpenStreetMap node ids.

weights=g.weights: the matrix that the edge weights are extracted from. Defaults to g.weights.

Return

Vector{W}: Array of edge weights, distances are in km, time is in hours.

LightOSM.astar — Function

astar(g::AbstractGraph{U},
      weights::AbstractMatrix{T},
      src::W,
      goal::W;
      heuristic::Function=h(u, v) =  0.0,
      cost_adjustment::Function=r(u, v) = 0.0
      ) where {T <: Real, U <: Integer, W <: Integer}

A* shortest path algorithm. Implemented with a min heap. Using a min heap is faster than using a priority queue given the sparse nature of OpenStreetMap data, i.e. vertices far outnumber edges.

Compared to jl, this version improves runtime, memory usage, has a flexible heuristic function, and accounts for OpenStreetMap turn restrictions through the cost_adjustment function.

Arguments

g::AbstractGraph{U}: Graphs abstract graph object.

weights::AbstractMatrix{T}: Edge weights matrix.

src::W: Source vertex.

goal::W: Goal vertex.

heuristic::Function=h(u, v) = 0.0: Heuristic cost function, takes a source and target vertex, default is 0.

cost_adjustment:::Function=r(u, v, parents) = 0.0: Optional cost adjustment function for use cases such as turn restrictions, takes a source and target vertex, defaults to 0.

Return

Vector{U}: Array veritces represeting shortest path between src to goal.

LightOSM.dijkstra — Function

dijkstra(g::AbstractGraph{U},
         weights::AbstractMatrix{T},
         src::W,
         goal::W;
         cost_adjustment::Function=r(u, v) = 0.0
         ) where {T <: Real, U <: Integer}

Dijkstra's shortest path algorithm with an early exit condition, is the same as astar with heuristic cost as 0.

Arguments

g::AbstractGraph{U}: Graphs abstract graph object.

weights::AbstractMatrix{T}: Edge weights matrix.

src::W: Source vertex.

goal::W: Goal vertex.

cost_adjustment:::Function=r(u, v, parents) = 0.0: Optional cost adjustment function for use cases such as turn restrictions, takes a source and target vertex, defaults to 0.

Return

Vector{U}: Array veritces represeting shortest path between src to goal.

dijkstra(g::AbstractGraph{U},
         weights::AbstractMatrix{T},
         src::W;
         cost_adjustment::Function=r(u, v) = 0.0
         ) where {T <: Real, U <: Integer, W <: Integer}

Dijkstra's shortest path algorithm, implemented with a min heap. Using a min heap is faster than using a priority queue given the sparse nature of OpenStreetMap data, i.e. vertices far outnumber edges.

This dispatch returns full set of parents or the dijkstra state given a source vertex, i.e. without and early exit condition of goal.

Arguments

g::AbstractGraph{U}: Graphs abstract graph object.

weights::AbstractMatrix{T}: Edge weights matrix.

src::W: Source vertex.

cost_adjustment:::Function=r(u, v, parents) = 0.0: Optional cost adjustment function for use cases such as turn restrictions, takes a source and target vertex, defaults to 0.

Return

Vector{U}: Array parent veritces from which the shortest path can be extracted.